requirements.txt
django==2.1
psycopg2-binary==2.7.5
-e git://github.com/mingchen/django-cas-ng.git#egg=django-cas-ng
django-tenants==2.0.0
unittest-xml-reporting==2.2.0
ddtrace==0.12.1
raven==6.9.0
djangorestframework==3.8.2
markdown==2.6.11
django-filter==2.0.0